requirements.txt
mysql-connector==2.2.9
mysql-connector-python==8.0.23
protobuf==3.14.0
PyQt6==6.0.1
PyQt6-Qt==6.0.1
PyQt6-sip==13.0.1
PySide6==6.0.1
shiboken6==6.0.1
six==1.15.0
